+++ src/components/StatusBadge/StatusBadge.tsx
@@ -6,7 +6,6 @@ export const statusBadgeFontSizeArray = [
   "desktop.body.body2",
   "desktop.caption",
 ] as const;
-export type StatusBadgeFontSizes = typeof statusBadgeFontSizeArray[number];
 
 export const statusBadgeLevelArray = ["low", "medium", "high"] as const;
 export type StatusBadgeLevels = typeof statusBadgeLevelArray[number];
@@ -50,10 +49,9 @@ export const StatusBadge: ChakraComponent<
   forwardRef<HTMLDivElement, React.PropsWithChildren<StatusBadgeProps>>(
     (props, ref?) => {
       const { children, className, id, level, type, ...rest } = props;
-      const labelFontSize = rest["fontSize"] || "desktop.body.body2";
       const finalVariant = level ? level : type ? type : "low";
       const styles = useStyleConfig("StatusBadge", {
-        labelFontSize,
+        labelFontSize: rest["fontSize"],
         variant: finalVariant,
       });